Fixed conflict issues due to the upgrade to the way that pirate values are stored and made test more robust using GameManager.dispose that was added before the old branch was created

+++ core/src/main/java/com/mygdx/game/Entitys/CannonBall.java
@@ -115,7 +115,7 @@ public class CannonBall extends Entity implements CollisionCallBack {
      * Added for Assessment 2
      * @return the damage dealt by the unit shooting this cannonball
      */
-    public int getAttackDmg() {
+    public Float getAttackDmg() {
         return shooter.getComponent(Pirate.class).getAttackDmg();
     }
